refactor(gguf): split metadata out + use macros
